manu conflicts fix
