+++ plugins/chat/config/routes.rb
@@ -17,6 +17,7 @@ Chat::Engine.routes.draw do
     post "/channels/:channel_id/memberships/me" => "channels_current_user_membership#create"
     put "/channels/:channel_id/notifications-settings/me" =>
           "channels_current_user_notifications_settings#update"
+    put "/tracking/read/me" => "tracking#read"
 
     # Category chatables controller hints. Only used by staff members, we don't want to leak category permissions.
     get "/category-chatables/:id/permissions" => "category_chatables#permissions",
